fix(pr): split modal in smaller components
